feat(layerslayer): package library and add tools link
