+++ layerslayer/__init__.py
@@ -1,4 +1,5 @@
 from .utils import parse_image_ref, registry_base_url, human_readable_size, load_token, save_token
+from .client import DockerRegistryClient, gather_layers_info, get_manifest, get_manifest_digest, list_layer_files
 
 __all__ = [
     "parse_image_ref",
@@ -6,4 +7,9 @@ __all__ = [
     "human_readable_size",
     "load_token",
     "save_token",
+    "DockerRegistryClient",
+    "gather_layers_info",
+    "get_manifest",
+    "get_manifest_digest",
+    "list_layer_files",
 ]
